+++ src/compiler/app-core/build-conditionals.ts
@@ -115,6 +115,7 @@ function getModuleImports(moduleMap: d.ModuleMap, filePath: string, importedModu
 export function updateBuildConditionals(config: d.Config, b: d.Build) {
   b.isDebug = (config.logLevel === 'debug');
   b.isDev = !!config.devMode;
+  b.devTools = b.isDev;
   b.lifecycleDOMEvents = !!(b.isDebug || config._isTesting || config._lifecycleDOMEvents);
   b.profile = !!(config.profile);
   b.hotModuleReplacement = !!(config.devMode && config.devServer && config.devServer.reloadStrategy === 'hmr' && !config._isTesting);