+++ src/compiler/component-lazy/generate-esm-browser.ts
@@ -4,14 +4,13 @@ import { generateLazyModules } from '../component-lazy/generate-lazy-module';
 import { OutputOptions, RollupBuild } from 'rollup';
 import { getDynamicImportFunction } from '@utils';
 
-export async function generateEsmBrowser(config: d.Config, compilerCtx: d.CompilerCtx, buildCtx: d.BuildCtx, build: d.Build, rollupBuild: RollupBuild, outputTargets: d.OutputTargetDistLazy[]) {
+export async function generateEsmBrowser(config: d.Config, compilerCtx: d.CompilerCtx, buildCtx: d.BuildCtx, rollupBuild: RollupBuild, outputTargets: d.OutputTargetDistLazy[]) {
   const esmOutputs = outputTargets.filter(o => !!o.esmDir && !!o.isBrowserBuild);
-  const isProd = !build.isDev;
   if (esmOutputs.length) {
     const esmOpts: OutputOptions = {
       format: 'esm',
       entryFileNames: '[name].esm.js',
-      chunkFileNames: isProd ? 'p-[hash].js' : '[name]-[hash].js',
+      chunkFileNames: config.hashFileNames ? 'p-[hash].js' : '[name]-[hash].js',
       preferConst: true,
       // This is needed until Firefox 67, which ships native dynamic imports
       dynamicImportFunction: getDynamicImportFunction(config.fsNamespace)