+++ src/runtime/bootstrap-lazy.ts
@@ -9,13 +9,15 @@ import { doc, getHostRef, plt, registerHost, supportsShadowDom, win } from '@pla
 import { hmrStart } from './hmr-component';
 import { HYDRATE_ID, PLATFORM_FLAGS, PROXY_FLAGS } from './runtime-constants';
 import { appDidLoad, forceUpdate } from './update-component';
-import { createTime } from './profile';
+import { createTime, installDevTools } from './profile';
 
 
 export const bootstrapLazy = (lazyBundles: d.LazyBundlesRuntimeData, options: d.CustomElementsDefineOptions = {}) => {
-  // if (BUILD.profile) {
-  //   performance.mark('st:app:start');
-  // }
+  if (BUILD.profile) {
+    performance.mark('st:app:start');
+  }
+  installDevTools();
+
   const endBootstrap = createTime('bootstrapLazy');
   const cmpTags: string[] = [];
   const exclude = options.exclude || [];
@@ -23,8 +25,8 @@ export const bootstrapLazy = (lazyBundles: d.LazyBundlesRuntimeData, options: d.
   const customElements = win.customElements;
   const y = /*@__PURE__*/head.querySelector('meta[charset]');
   const visibilityStyle = /*@__PURE__*/doc.createElement('style');
-  let appLoadFallback: any;
   const deferredConnectedCallbacks: {connectedCallback: () => void}[] = [];
+  let appLoadFallback: any;
   let isBootstrapping = true;
 
   Object.assign(plt, options);
@@ -39,16 +41,11 @@ export const bootstrapLazy = (lazyBundles: d.LazyBundlesRuntimeData, options: d.
   }
   if (BUILD.hydrateClientSide && BUILD.shadowDom) {
     const styles = doc.querySelectorAll('style[s-id]');
-    let globalStyles = '';
-    let i = 0;
-    for (; i < styles.length; i++) {
-      globalStyles += styles[i].innerHTML;
-    }
-    for (i = 0; i < styles.length; i++) {
+    for (let i = 0; i < styles.length; i++) {
       const styleElm = styles[i];
       registerStyle(
         styleElm.getAttribute(HYDRATE_ID),
-        globalStyles + convertScopedToShadow(styleElm.innerHTML),
+        convertScopedToShadow(styleElm.innerHTML),
         true,
       );
     }
@@ -103,15 +100,14 @@ export const bootstrapLazy = (lazyBundles: d.LazyBundlesRuntimeData, options: d.
         }
 
         connectedCallback() {
+          if (appLoadFallback) {
+            clearTimeout(appLoadFallback);
+            appLoadFallback = null;
+          }
           if (isBootstrapping) {
             // connectedCallback will be processed once all components have been registered
             deferredConnectedCallbacks.push(this);
           } else {
-            if (appLoadFallback) {
-              clearTimeout(appLoadFallback);
-              appLoadFallback = null;
-            }
-
             plt.jmp(() => connectedCallback(this, cmpMeta));
           }
         }
@@ -137,6 +133,7 @@ export const bootstrapLazy = (lazyBundles: d.LazyBundlesRuntimeData, options: d.
       cmpMeta.$lazyBundleIds$ = lazyBundle[0];
 
       if (!exclude.includes(tagName) && !customElements.get(tagName)) {
+        cmpTags.push(tagName);
         customElements.define(
           tagName,
           proxyComponent(HostElement as any, cmpMeta, PROXY_FLAGS.isElementConstructor) as any
@@ -144,16 +141,18 @@ export const bootstrapLazy = (lazyBundles: d.LazyBundlesRuntimeData, options: d.
       }
     }));
 
-  // Process deferred connectedCallbacks now all components have been registered
-  isBootstrapping = false;
-  deferredConnectedCallbacks.forEach(host => host.connectedCallback());
-
   // visibilityStyle.innerHTML = cmpTags.map(t => `${t}:not(.hydrated)`) + '{display:none}';
   visibilityStyle.innerHTML = cmpTags + '{visibility:hidden}.hydrated{visibility:inherit}';
   visibilityStyle.setAttribute('data-styles', '');
   head.insertBefore(visibilityStyle, y ? y.nextSibling : head.firstChild);
 
+  // Process deferred connectedCallbacks now all components have been registered
+  isBootstrapping = false;
+  if (deferredConnectedCallbacks.length > 0) {
+    deferredConnectedCallbacks.forEach(host => host.connectedCallback());
+  } else {
+    plt.jmp(() => appLoadFallback = setTimeout(appDidLoad, 30, 'timeout'));
+  }
   // Fallback appLoad event
-  plt.jmp(() => appLoadFallback = setTimeout(appDidLoad, 30));
   endBootstrap();
 };