+++ src/runtime/set-value.ts
@@ -4,6 +4,7 @@ import { consoleError, getHostRef } from '@platform';
 import { HOST_FLAGS } from '@utils';
 import { parsePropertyValue } from './parse-property-value';
 import { scheduleUpdate } from './update-component';
+import { STENCIL_DEV_MODE } from './profile';
 
 
 export const getValue = (ref: d.RuntimeRef, propName: string) =>
@@ -23,13 +24,24 @@ export const setValue = (ref: d.RuntimeRef, propName: string, newVal: any, cmpMe
     // set our new value!
     hostRef.$instanceValues$.set(propName, newVal);
 
-    if (BUILD.isDev && hostRef.$flags$ & HOST_FLAGS.dev_stateMutation) {
-      console.warn(
-        `[STENCIL-DEV-MODE] The state/prop "${propName}" changed during rendering, triggering another re-render. This is not a recommended pattern since it can lead to infinite-loops and other bugs.`,
-        'Element', elm,
-        'New value', newVal,
-        'Old value', oldVal
-      );
+    if (BUILD.isDev) {
+      if (hostRef.$flags$ & HOST_FLAGS.devOnRender) {
+        console.warn(
+          ...STENCIL_DEV_MODE,
+          `The state/prop "${propName}" changed during rendering. This can potentially lead to infinite-loops and other bugs.`,
+          '\nElement', elm,
+          '\nNew value', newVal,
+          '\nOld value', oldVal
+        );
+      } else if (hostRef.$flags$ & HOST_FLAGS.devOnDidLoad) {
+        console.debug(
+          ...STENCIL_DEV_MODE,
+          `The state/prop "${propName}" changed during "componentDidLoad()", this triggers extra re-renders, try to setup on "componentWillRender()"`,
+          '\nElement', elm,
+          '\nNew value', newVal,
+          '\nOld value', oldVal
+        );
+      }
     }
 
     if (!BUILD.lazyLoad || instance) {