+++ src/runtime/test/render-vdom.spec.tsx
@@ -584,6 +584,7 @@ describe('render-vdom', () => {
               <input type='password' />
               <input type='email' />
               <input type='date' />
+              <input list='my-list' />
             </Host>
           );
         }
@@ -595,13 +596,14 @@ describe('render-vdom', () => {
       });
       expect(root).toEqualHtml(`
         <cmp-a>
-          <button type='button'></button>
-          <button type='submit'></button>
+          <button type=\"button\"></button>
+          <button type=\"submit\"></button>
           <input type=\"text\" value=\"\">
           <input type=\"number\">
           <input type=\"password\">
           <input type=\"email\">
           <input type=\"date\">
+          <input list=\"my-list\" />
         </cmp-a>
       `);
     });