+++ test/prerender-shadow/stencil.config.ts
@@ -3,7 +3,7 @@ import { Config } from '../../dist';
 export const config: Config = {
 
   hashFileNames: false,
-
+  minifyJs: false,
   outputTargets: [
     {
       type: 'www',